Removed particles and implemented fake animation for chest opening
What the client wanted was not particles for the chests, he wanted a fake open/close sound and animation for the chests, since when canceling the chest opening event, even when displaying a fake inventory with `player.openInventory`, the animation doesn't occurs normally.

Fixed minor typo in command `unmark` text, removed unused commented code in ChestRepo.